update dependencies, and run forge fmt

+++ src/DsSort.sol
@@ -1,29 +1,25 @@
 pragma solidity >= 0.8.0;
 
 contract DsSort {
-    function sort(uint[] memory data) public pure returns (uint[] memory) {
+    function sort(uint256[] memory data) public pure returns (uint256[] memory) {
         unchecked {
             quickSort(data, 0, data.length - 1);
         }
         return data;
     }
 
-    function safeSub(uint a) public pure returns (uint) {
+    function safeSub(uint256 a) public pure returns (uint256) {
         unchecked {
             if (a == 0) return a;
             else return a - 1;
         }
     }
 
-    function quickSort(
-        uint[] memory arr,
-        uint left,
-        uint right
-    ) internal pure {
+    function quickSort(uint256[] memory arr, uint256 left, uint256 right) internal pure {
         unchecked {
-            uint pivot = arr[left + ((right - left) / 2)];
-            uint i = left;
-            uint j = right;
+            uint256 pivot = arr[left + ((right - left) / 2)];
+            uint256 i = left;
+            uint256 j = right;
             // if (i == j) return;
             while (true) {
                 while (arr[i] < pivot) i++;
